finish cleaning args.py
